bookmarks/requirements.txt
amqp==5.0.5
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.7
django-celery-results==2.0.1
django-cors-headers==3.7.0
django-filter==2.4.0
django-sendgrid-v5==0.9.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
factory-boy==3.2.0
Faker==6.5.0
future==0.18.2
http-status-code==0.0.8
idna==2.10
kombu==5.0.2
Markdown==3.3.4
parameterized==0.8.1
Pillow==8.1.2
prompt-toolkit==3.0.17
psycopg2==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.4
python-http-client==3.3.2
pytz==2021.1
redis==3.5.3
requests==2.25.1
sendgrid==6.6.0
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
stripe==2.56.0
text-unidecode==1.3
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
